requirements.txt
annotated-types
anyio==4.4.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.1
click==8.1.7
cssselect2==0.7.0
diff-match-patch==20230430
Django==4.2
django-import-export==4.1.1
django-jazzmin==3.0.0
django-solo==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.99.0
fonttools==4.53.1
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
idna==3.8
openpyxl==3.1.5
pillow==10.4.0
pycparser==2.22
pydantic==1.10.18
pydantic_core==2.20.1
pydyf==0.11.0
pyphen==0.16.0
python-dotenv==1.0.1
python-telegram-bot==20.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.27.0
tablib==3.5.0
tinycss2==1.3.0
typing_extensions==4.12.2
uvicorn==0.22.0
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3
PyPDF2
django-mathfilters==1.0.0
sentry-sdk
